fix: garden 삭제 rest docs 오류 수정
